add facets (aggregations) to the search response
Main thing of this commit is aggregations from elasticsearch are
queried and filled into the response. For the start only for
creators. With that it should be possible to use facets in the
frontend.
Additionally some small changes: read subtitles and
PlaceOfPublish from the elasticsearch-results. use size (limit)
and from (offset).