probably slow implementation for offset and limit
this probably solves the problem to set limit and offset for the
search. But this is puerely implemented on the java-side and the query
fetches all possible data and cuts with limit and offset after
retrieval. Possibly this is to slow when many datasets are inside
elasticsearch and has to be changed.