Simplify and improve the facet-search
Currently the olad search can still be used, so that it is possible to
compare the new to the old version.
To remove it ElasticQueryHelper.java should be deleted, every not used
Method from ElasticResponseHelper.java and
ElasticSearchService.facetSearch()

The old version searches for the searchterm in all documents of all
indices. This way duplicates for a single pid/workspace is possible.
Therefore the results are grouped by pid. This can lead to problems
regarding facets because not all documents have all the information.
More details in Jira: OLAHDS-445.

The new version only searches in log-entries having isFirst = true so
that only one hit per pid/workspace is possible. This simplifies the
search but not all information is searchable any more because not all
entries are queried any more. It is a trade off, but the results are
better especially the facets. Also this way it would be possible to
implement sorting and I think this must be faster because less
aggregations are used (not verified though).